load_data_to_pinecone/requirements.txt
pinecone
google
google-api-python-client
grpcio
protoc_gen_openapiv2
uuid
python-dotenv